+++ data_access/handlers/__class_abstract_handler.py
@@ -3,8 +3,9 @@ from pathlib import Path
 
 
 class AbstractHandler(ABC):
-    def __init__(self, link, work_dir=None):
-        self.link = Path(link)
+    def __init__(self, link, link_folder, work_dir=None):
+        self.link = link
+        self.link_folder = link_folder
         self.work_dir = Path(work_dir) if work_dir else None
 
     @abstractmethod
@@ -14,3 +15,4 @@ class AbstractHandler(ABC):
     @abstractmethod
     def put(self, data):
         pass
+
