+++ data_access/handlers/__class_feather_handler.py
@@ -3,8 +3,8 @@ from data_access.handlers.__class_abstract_handler import AbstractHandler
 
 
 class Feather(AbstractHandler):
-    def __init__(self, link, work_dir=None):
-        super().__init__(link, work_dir=work_dir)
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
     def get(self):
         return pd.read_feather(self.link)
