+++ data_access/handlers/__class_json_handler.py
@@ -3,8 +3,8 @@ from data_access.handlers.__class_abstract_handler import AbstractHandler
 
 
 class Json(AbstractHandler):
-    def __init__(self, link, work_dir=None):
-        super().__init__(link, work_dir=work_dir)
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
     def get(self):
         with open(self.link) as json_file:
