--- data_access/tests/work_dir/aaaBBB.txt
@@ -1,1 +1,0 @@
-it is a file for testing
